Edited mygroup page to show the profiles only if there are verba matches.
