+++ lib/screens/myGroup.dart
@@ -7,6 +7,7 @@ import 'package:verbatim_frontend/screens/addFriend.dart';
 import 'package:verbatim_frontend/screens/customChallenge.dart';
 import 'package:verbatim_frontend/screens/groupChallenge.dart';
 import 'package:verbatim_frontend/widgets/firebase_download_image.dart';
+import 'package:verbatim_frontend/widgets/statsContent.dart';
 import 'sideBar.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:verbatim_frontend/widgets/size.dart';
@@ -844,398 +845,5 @@ class _MyGroupState extends State<myGroup> with SingleTickerProviderStateMixin {
               ]))),
       drawer: const SideBar(),
     ));
-  }
-}
-
-class StatsContent extends StatelessWidget {
-  List<String> verbaMatchStatsContent;
-  double groupRating;
-
-  StatsContent({
-    super.key,
-    required this.verbaMatchStatsContent,
-    required this.groupRating,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    bool isEmpty = false;
-    print("this is grouprating in groups stats content $groupRating");
-    if (verbaMatchStatsContent.isEmpty) {
-      isEmpty = true;
-      verbaMatchStatsContent = ["i", "amempty"];
-    }
-    String verb1 = verbaMatchStatsContent[0];
-    String verb2 = verbaMatchStatsContent[1];
-
-    print(
-        "this is verbamatch in groups stats content $String verb1 = verbaMatchStatsContent[0];");
-    return SingleChildScrollView(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          SizedBox(
-            // color: Colors.white,
-            height: 200,
-            child: Padding(
-              padding: const EdgeInsets.all(16.0),
-              child: DonutChart(
-                  groupSimilarity: groupRating, title: 'Group Power Score'),
-            ),
-          ),
-          SizedBox(height: 15.v),
-          Visibility(
-            visible: !isEmpty,
-            child: Container(
-              child: const Center(
-                child: Text(
-                  'Most Similar: ',
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            ),
-          ),
-          Visibility(
-              visible: isEmpty,
-              child: Container(
-                child: Center(
-                  child: RichText(
-                    text: const TextSpan(
-                      children: [
-                        TextSpan(
-                          text: 'No ',
-                          style: TextStyle(
-                            color: Colors.black,
-                            fontSize: 24,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        TextSpan(
-                          text: 'Verba-',
-                          style: TextStyle(
-                            color: Colors.orange,
-                            fontSize: 24,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        TextSpan(
-                          text: 'Matches...',
-                          style: TextStyle(
-                            color: Colors.black,
-                            fontSize: 24,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              )),
-          Visibility(
-            visible: verbaMatchStatsContent.isEmpty,
-            child: const SizedBox(
-              height: 20,
-            ),
-          ),
-          Visibility(
-            visible: isEmpty,
-            child: Container(
-              child: const Center(
-                child: Text(
-                  'Play more challenges to match!',
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            ),
-          ),
-          Visibility(
-            visible: !isEmpty,
-            child: Container(
-              child: RichText(
-                text: const TextSpan(
-                  children: [
-                    TextSpan(
-                      text: 'Verba',
-                      style: TextStyle(
-                        color: Colors.orange,
-                        fontSize: 24,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                    TextSpan(
-                      text: 'Match!',
-                      style: TextStyle(
-                        color: Colors.black,
-                        fontSize: 24,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ),
-          /*
-          Visibility(
-            visible: !isEmpty,
-            child: Container(
-              child: Center(
-                child: Text(
-                  'hardcoded% Similarity',
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            ),
-          ),
-          */
-          Visibility(
-            visible: !isEmpty,
-            child: SizedBox(height: 10.v),
-          ),
-          Visibility(
-            visible: !isEmpty,
-            child: Center(
-              child: SizedBox(
-                width: 170,
-                height: 60,
-                child: Stack(
-                  children: [
-                    for (int i = 0; i < 2; i++)
-                      Positioned(
-                        top: 0,
-                        left: 35.0 + (i * 50),
-                        child: Image.asset(
-                          'assets/Ellipse ${41 + i}.png',
-                          height: 60,
-                        ),
-                      ),
-                  ],
-                ),
-              ),
-            ),
-          ),
-          Visibility(
-            visible: !isEmpty,
-            child: SizedBox(height: 15.v),
-          ),
-          Visibility(
-            visible: !isEmpty,
-            child: Container(
-              child: Center(
-                child: Text(
-                  '$verb1 and $verb2',
-                  style: const TextStyle(
-                    fontSize: 14,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            ),
-          ),
-          SizedBox(height: 50.v)
-        ],
-      ),
-    );
-  }
-}
-
-class DonutChart extends StatefulWidget {
-  final double groupSimilarity;
-  final String title;
-
-  const DonutChart({
-    Key? key,
-    required this.groupSimilarity,
-    required this.title,
-  }) : super(key: key);
-
-  @override
-  State<DonutChart> createState() => _DonutChartState();
-}
-
-class _DonutChartState extends State<DonutChart> {
-  Future<void> _showPopup(BuildContext context) async {
-    return showDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          contentPadding: EdgeInsets.zero,
-          content: Container(
-              width: 160,
-              height: 145,
-              clipBehavior: Clip.antiAlias,
-              decoration: BoxDecoration(
-                color: Colors.white,
-                borderRadius: BorderRadius.circular(20),
-                boxShadow: [
-                  BoxShadow(
-                    color: const Color(0xFF9E503C).withOpacity(0.5),
-                    blurRadius: 4,
-                    offset: const Offset(2, 3),
-                  )
-                ],
-              ),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  const SizedBox(height: 5),
-                  Padding(
-                    padding: const EdgeInsets.all(6),
-                    child: RichText(
-                      textAlign: TextAlign.center,
-                      text: const TextSpan(
-                        children: [
-                          TextSpan(
-                            text: 'Your score increases when you:\n\n',
-                            style: TextStyle(
-                              color: Colors.black,
-                              fontSize: 19,
-                              fontWeight: FontWeight.bold,
-                            ),
-                          ),
-                          // if groupname is a certain length, make it a new line
-                          TextSpan(
-                            text:
-                                'Verbatim, Build Streaks, and Play Challenges!',
-                            style: TextStyle(
-                              color: Colors.orange,
-                              fontSize: 21,
-                              fontWeight: FontWeight.w900,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  )
-                ],
-              )),
-        );
-      },
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    Color calculateColor(double similarity) {
-      int score = similarity ~/ 2;
-
-      return Color.fromARGB(255, 250, 192 + score, 94 + score);
-    }
-
-    double sim = widget.groupSimilarity;
-    int simint = sim as int;
-
-    return Scaffold(
-      backgroundColor: Colors.transparent,
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-            Text(
-              widget.title,
-              style: const TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            const SizedBox(width: 8),
-            InkWell(
-              onTap: () => _showPopup(context),
-              child: const Icon(
-                Icons.help_outline,
-                color: Color(0xFFE76F51),
-              ),
-            )
-          ]),
-          SizedBox(height: 17.v),
-          SizedBox(
-            height: 125,
-            child: Stack(
-              children: [
-                PieChart(
-                  //set the values of offset
-
-                  PieChartData(
-                    startDegreeOffset: 250,
-                    sectionsSpace: 0,
-                    centerSpaceRadius: 50,
-                    sections: [
-                      PieChartSectionData(
-                        value: widget.groupSimilarity,
-                        color: const Color(0xFFE76F51),
-                        radius: 25,
-                        showTitle: false,
-                      ),
-                      PieChartSectionData(
-                        value: 100 - widget.groupSimilarity,
-                        color: calculateColor(widget.groupSimilarity),
-                        radius: 16,
-                        showTitle: false,
-                      ),
-                    ],
-                  ),
-                ),
-                // text inside chart
-                Positioned.fill(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Container(
-                        height: 80,
-                        width: 80,
-                        decoration: const BoxDecoration(
-                          color: Colors.white,
-                          shape: BoxShape.circle,
-                          boxShadow: [
-                            BoxShadow(
-                                color: Color.fromARGB(255, 255, 243, 238),
-                                blurRadius: 10.0,
-                                spreadRadius: 10.0,
-                                offset: Offset(3, 3)),
-                          ],
-                        ),
-                        child: Align(
-                          alignment: Alignment.center,
-                          child: Column(
-                            mainAxisSize: MainAxisSize.min,
-                            children: [
-                              Text(
-                                simint.toStringAsFixed(2),
-                                style: const TextStyle(
-                                  fontSize: 10,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                              const Text(
-                                "Rating",
-                                style: TextStyle(
-                                  fontSize: 10,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                      )
-                    ],
-                  ),
-                )
-              ],
-            ),
-          )
-        ],
-      ),
-    );
   }
 }
