+++ lib/widgets/donutChart.dart
@@ -1,0 +1,195 @@
+import 'package:fl_chart/fl_chart.dart';
+import 'package:flutter/material.dart';
+import 'package:verbatim_frontend/widgets/size.dart';
+
+class DonutChart extends StatefulWidget {
+  final double groupSimilarity;
+  final String title;
+
+  const DonutChart({
+    Key? key,
+    required this.groupSimilarity,
+    required this.title,
+  }) : super(key: key);
+
+  @override
+  State<DonutChart> createState() => _DonutChartState();
+}
+
+class _DonutChartState extends State<DonutChart> {
+  Future<void> _showPopup(BuildContext context) async {
+    return showDialog<void>(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          contentPadding: EdgeInsets.zero,
+          content: Container(
+              width: 160,
+              height: 145,
+              clipBehavior: Clip.antiAlias,
+              decoration: BoxDecoration(
+                color: Colors.white,
+                borderRadius: BorderRadius.circular(20),
+                boxShadow: [
+                  BoxShadow(
+                    color: const Color(0xFF9E503C).withOpacity(0.5),
+                    blurRadius: 4,
+                    offset: const Offset(2, 3),
+                  )
+                ],
+              ),
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                mainAxisAlignment: MainAxisAlignment.center,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: [
+                  const SizedBox(height: 5),
+                  Padding(
+                    padding: const EdgeInsets.all(6),
+                    child: RichText(
+                      textAlign: TextAlign.center,
+                      text: const TextSpan(
+                        children: [
+                          TextSpan(
+                            text: 'Your score increases when you:\n\n',
+                            style: TextStyle(
+                              color: Colors.black,
+                              fontSize: 19,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                          // if groupname is a certain length, make it a new line
+                          TextSpan(
+                            text:
+                                'Verbatim, Build Streaks, and Play Challenges!',
+                            style: TextStyle(
+                              color: Colors.orange,
+                              fontSize: 21,
+                              fontWeight: FontWeight.w900,
+                            ),
+                          ),
+                        ],
+                      ),
+                    ),
+                  )
+                ],
+              )),
+        );
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    Color calculateColor(double similarity) {
+      int score = similarity ~/ 2;
+
+      return Color.fromARGB(255, 250, 192 + score, 94 + score);
+    }
+
+    double sim = widget.groupSimilarity;
+    int simint = sim as int;
+
+    return Scaffold(
+      backgroundColor: Colors.transparent,
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Row(mainAxisAlignment: MainAxisAlignment.center, children: [
+            Text(
+              widget.title,
+              style: const TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+            const SizedBox(width: 8),
+            InkWell(
+              onTap: () => _showPopup(context),
+              child: const Icon(
+                Icons.help_outline,
+                color: Color(0xFFE76F51),
+              ),
+            )
+          ]),
+          SizedBox(height: 17.v),
+          SizedBox(
+            height: 125,
+            child: Stack(
+              children: [
+                PieChart(
+                  //set the values of offset
+
+                  PieChartData(
+                    startDegreeOffset: 250,
+                    sectionsSpace: 0,
+                    centerSpaceRadius: 50,
+                    sections: [
+                      PieChartSectionData(
+                        value: widget.groupSimilarity,
+                        color: const Color(0xFFE76F51),
+                        radius: 25,
+                        showTitle: false,
+                      ),
+                      PieChartSectionData(
+                        value: 100 - widget.groupSimilarity,
+                        color: calculateColor(widget.groupSimilarity),
+                        radius: 16,
+                        showTitle: false,
+                      ),
+                    ],
+                  ),
+                ),
+                // text inside chart
+                Positioned.fill(
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Container(
+                        height: 80,
+                        width: 80,
+                        decoration: const BoxDecoration(
+                          color: Colors.white,
+                          shape: BoxShape.circle,
+                          boxShadow: [
+                            BoxShadow(
+                                color: Color.fromARGB(255, 255, 243, 238),
+                                blurRadius: 10.0,
+                                spreadRadius: 10.0,
+                                offset: Offset(3, 3)),
+                          ],
+                        ),
+                        child: Align(
+                          alignment: Alignment.center,
+                          child: Column(
+                            mainAxisSize: MainAxisSize.min,
+                            children: [
+                              Text(
+                                simint.toStringAsFixed(2),
+                                style: const TextStyle(
+                                  fontSize: 10,
+                                  fontWeight: FontWeight.bold,
+                                ),
+                              ),
+                              const Text(
+                                "Rating",
+                                style: TextStyle(
+                                  fontSize: 10,
+                                  fontWeight: FontWeight.bold,
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                      )
+                    ],
+                  ),
+                )
+              ],
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
